refactor(Movie): 상영 일자 및 상영 시간 반환 로직 구현
